最终版本提交

+++ src/main/java/com/wusiq/controller/UserController.java
@@ -13,8 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import java.util.List;
@@ -32,7 +31,7 @@ public class UserController {
 
     /*增*/
     @RequestMapping(value = "/addRow.do",method = RequestMethod.GET)
-    public ModelAndView addRow(@Validated ReqAddRowDto reqAddRowDto, BindingResult result){
+    public ModelAndView addRow(@Validated @RequestBody ReqAddRowDto reqAddRowDto, BindingResult result){
         LOGGER.info("UserController.addRow.start.param:{}", JSON.toJSONString(reqAddRowDto));
         ModelAndView modelAndView = new ModelAndView();
         if (result.hasErrors()){
@@ -41,7 +40,7 @@ public class UserController {
             return modelAndView;
         }
 
-    UserEntity userEntity = new UserEntity();
+        UserEntity userEntity = new UserEntity();
         userEntity.setUserNmae(reqAddRowDto.getUserName());
         userEntity.setUserAge(reqAddRowDto.getUserAge());
         boolean bool = userService.addRow(userEntity);
@@ -55,18 +54,26 @@ public class UserController {
     }
 
     /*删*/
-    @RequestMapping(value = "/deleteRow.json",method = RequestMethod.GET)
-    public String deleteRow(ReqDeleteRowDto reqDeleteRowDto){
-        LOGGER.info("UserController.deleteRow.start...");
+    @ResponseBody
+    @RequestMapping(value = "/deleteRow.do",method = RequestMethod.GET)
+    public String deleteRow(@Validated @RequestBody ReqDeleteRowDto reqDeleteRowDto, BindingResult result){
+        LOGGER.info("UserController.deleteRow.start.param:{}", JSON.toJSONString(reqDeleteRowDto));
+        if (result.hasErrors()){
+            return "fail";
+        }
         boolean bool = userService.deleteRow(reqDeleteRowDto.getId());
         LOGGER.info("UserController.deleteRow.end.result:{}", bool);
         return bool?"success":"fail";
     }
 
     /*改*/
-    @RequestMapping(value = "/updateRow.json",method = RequestMethod.GET)
-    public String updateRow(ReqUpdateRowDto reqUpdateRowDto){
-        LOGGER.info("UserController.updateRow.start...");
+    @ResponseBody
+    @RequestMapping(value = "/updateRow.do",method = RequestMethod.GET)
+    public String updateRow(@Validated @RequestBody ReqUpdateRowDto reqUpdateRowDto, BindingResult result){
+        LOGGER.info("UserController.updateRow.start.param:{}", JSON.toJSONString(reqUpdateRowDto));
+        if (result.hasErrors()){
+            return "fail";
+        }
         UserEntity userEntity = new UserEntity();
         userEntity.setId(reqUpdateRowDto.getId());
         userEntity.setUserNmae(reqUpdateRowDto.getUserNmae());
@@ -77,15 +84,17 @@ public class UserController {
     }
 
     /*查*/
+    @ResponseBody
     @RequestMapping(value = "/queryRow.json",method = RequestMethod.GET)
-    public String queryRow(ReqQueryRowDto reqQueryRowDto){
-        LOGGER.info("UserController.queryRow.start...");
+    public String queryRow(@Validated @RequestBody ReqQueryRowDto reqQueryRowDto, BindingResult result){
+        LOGGER.info("UserController.queryRow.start.param:{}", JSON.toJSONString(reqQueryRowDto));
         UserEntity userEntity = userService.queryRow(reqQueryRowDto.getId());
         LOGGER.info("UserController.queryRow.end.result:{}", JSON.toJSONString(userEntity));
         return JSON.toJSONString(userEntity);
     }
 
-   /*查list*/
+    /*查list*/
+    @ResponseBody
     @RequestMapping(value = "/queryRowList.json",method = RequestMethod.GET)
     public String queryRowList(){
         LOGGER.info("UserController.queryRowList.start...");
